Confirmación de borrado

+++ backend/src/main/webapp/js/listado.js
@@ -4,62 +4,62 @@ let form, table, inputId, inputNombre, inputFecha;
 
 const URL = 'http://localhost:8080/backend/api/v1/usuarios';
 
-window.addEventListener('DOMContentLoaded', async function () {
-    form = document.querySelector('form');
-    table = document.querySelector('table');
+window.addEventListener('DOMContentLoaded', async function() {
+	form = document.querySelector('form');
+	table = document.querySelector('table');
 
-    inputId = document.querySelector('#id');
-    inputNombre = document.querySelector('#nombre');
-    inputFecha = document.querySelector('#fecha');
+	inputId = document.querySelector('#id');
+	inputNombre = document.querySelector('#nombre');
+	inputFecha = document.querySelector('#fecha');
 
-    form.style.display = 'none';
+	form.style.display = 'none';
 
-    form.addEventListener('submit', guardar);
+	form.addEventListener('submit', guardar);
 
-    await recargarTabla();
+	await recargarTabla();
 });
 
 async function guardar(e) {
-    e.preventDefault();
+	e.preventDefault();
 
-    table.style.display = 'table';
-    form.style.display = 'none';
+	table.style.display = 'table';
+	form.style.display = 'none';
 
-    if (inputId.value) {
+	if (inputId.value) {
 		const usuario = { id: inputId.value, nombre: inputNombre.value, fechaUltimaConexion: inputFecha.value };
-		
-        await fetch(`${URL}/${inputId.value}`, {
-            method: 'PUT',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(usuario)
-        });
-    } else {
-		const usuario = { nombre: inputNombre.value, fechaUltimaConexion: inputFecha.value };
-		
-        await fetch(URL, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify(usuario)
-        });
-    }
 
-    recargarTabla();
+		await fetch(`${URL}/${inputId.value}`, {
+			method: 'PUT',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify(usuario)
+		});
+	} else {
+		const usuario = { nombre: inputNombre.value, fechaUltimaConexion: inputFecha.value };
+
+		await fetch(URL, {
+			method: 'POST',
+			headers: {
+				'Content-Type': 'application/json'
+			},
+			body: JSON.stringify(usuario)
+		});
+	}
+
+	recargarTabla();
 }
 async function recargarTabla() {
-    const respuesta = await fetch(URL);
-    const usuarios = await respuesta.json();
+	const respuesta = await fetch(URL);
+	const usuarios = await respuesta.json();
 
-    const tbody = document.querySelector('tbody');
+	const tbody = document.querySelector('tbody');
 
-    tbody.innerHTML = '';
+	tbody.innerHTML = '';
 
-    usuarios.forEach(usuario => {
-        const tr = document.createElement('tr');
-        tr.innerHTML = `
+	usuarios.forEach(usuario => {
+		const tr = document.createElement('tr');
+		tr.innerHTML = `
             <td>${usuario.id}</td>
             <td>${usuario.nombre}</td>
             <td>${usuario.fechaUltimaConexion}</td>
@@ -69,38 +69,42 @@ async function recargarTabla() {
             </td>
             `;
 
-        tbody.appendChild(tr);
-    });
+		tbody.appendChild(tr);
+	});
 }
 
 async function formulario(id) {
-    if(id) {
-	    const respuesta = await fetch(`${URL}/${id}`);
-	    
-	    console.log(respuesta);
-	    
-	    if(respuesta.status === 404) {
+	if (id) {
+		const respuesta = await fetch(`${URL}/${id}`);
+
+		console.log(respuesta);
+
+		if (respuesta.status === 404) {
 			await recargarTabla();
 			alert('No se ha encontrado el registro y se ha refrescado la tabla con sus valores actuales');
 			return;
 		}
-	    
-	    const usuario = await respuesta.json();
 
-        inputId.value = usuario.id;
-        inputNombre.value = usuario.nombre;
-        inputFecha.value = usuario.fechaUltimaConexion;
-    } else {
-        inputId.value = '';
-        inputNombre.value = '';
-        inputFecha.value = '';
-    }
-    
-    table.style.display = 'none';
-    form.style.display = 'block';
+		const usuario = await respuesta.json();
+
+		inputId.value = usuario.id;
+		inputNombre.value = usuario.nombre;
+		inputFecha.value = usuario.fechaUltimaConexion;
+	} else {
+		inputId.value = '';
+		inputNombre.value = '';
+		inputFecha.value = '';
+	}
+
+	table.style.display = 'none';
+	form.style.display = 'block';
 }
 
 async function borrar(id) {
-    await fetch(`${URL}/${id}`, { method: 'DELETE' });
-    recargarTabla();
+	const quiereBorrar = confirm(`¿Estás seguro de que quieres borrar el registro id ${id}?`);
+
+	if (quiereBorrar) {
+		await fetch(`${URL}/${id}`, { method: 'DELETE' });
+		recargarTabla();
+	}
 }
